backend-git/backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.115.6
flatbuffers==24.3.25
h11==0.14.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.0.2
onnxruntime==1.19.2
packaging==24.2
pillow==11.0.0
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
pyreadline3==3.5.4
python-multipart==0.0.19
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
typing_extensions==4.12.2
uvicorn==0.32.1
